// matrix-app-front/src/shared/enums/routes/card-payment-routes.enum.ts
export enum CardPaymentRoutesEnum {
  CardPaymentStack = 'CardPaymentStack',
  ACCOUNT_STATUS = 'AccountStatus',
  ADD_DEBIT_CARD_CULQUI = 'AddDebitCardCulqi',
  ADDING_PAYMENT_METHOD = 'AddingPaymentMethod',
  CARD_PAYMENT = 'CardPayment',
  LIST_ACCOUNT_STATEMENTS = 'ListAccountStatements',
  PAYMENT_ERROR = 'PaymentError',
  PAYMENT_LOADING = 'PaymentLoading',
  PAYMENT_SUCCESS = 'PaymentSuccess',
}